lib: Don't use EOS propagation for cleanup
It's unnecessary, unreliable, and convoluted. We'll never receive an EOS on any
of our receive/transmit streams since they're UDP-based. v4l2src won't give an
EOS without num-buffers set, and if it ends, it'll be due to an ERROR.

As a plus, when we call one_video_local_peer_stop, we know that everything has
been freed as soon as it returns.