lib: Use proxysrc and proxysink elements
This allows us to proxy events and queries between the two pipelines, which
allows us to not have to force caps everywhere, allows the pipelines to manage
themselves, etc. For instance, it allows us to remove the capsfilter and
videorate in the remote receive pipeline. This is important because unless it's
in drop-only mode, videorate adds latency of at least two frames.