lib: Implement negotiation_stop using special states
Convert OneVideoLocalState to a bitwise enum which allows us to have multiple
states at once. This allows us to signal failure at a specific stage of the
negotiation/call process and also store whether we're in a state that places us
as a negotiator or a negotiatee or neither.

Also, free remote peer objects as part of _stop instead of _stop_playback since
they're created before playing is begun.