lib: Send and recv RTCP receiver reports as well
This requires us to notify remotes about another pair of ports and fix the
transmit and remote receive pipelines to do the send/recv over the network using
the appropriate rtpbins.

We receive RTCP RR (receiver report) feedback from all the receivers on the same
ports. Rtpbin is able to differentiate them based on the SSRC in the packet
header.

Finally, we use the same port for receiving RRs that we use for sending SRs
(sender reports) to ease the addition of NAT traversal later. We do the same on
the receiver side with receiving SRs and sending RRs for the same reason.

In the process of doing all this, the wire format changed, some struct members
were renamed to be clearer, and so on.