lib, app: Don't force listening on a single interface
Also allow specifying "any" as an interface for one-video-app

This required a lot of changes and has changed the wire protocol significantly
but is quite necessary for us since originally each peer was identified by an
addr:port pair where the port was the TCP server port, and this pair had to be
sent by each peer.

However, this is messy when listening on multiple interfaces since you have to
connect to a remote first, then extract the local address used for that, and
then construct the message that identifies us to the remote. It's simpler and
possibly more future-proof to let the remote detect the address it's seeing us
as and just communicate the port that it should use to connect back to us.

In the future when dealing with NATs, hopefully we can find a way to directly
replace that with an allocated port on the NAT.

We now use a unique peer id that is generated at startup to identify each peer
in the pool. This peer id is generated from the host name, the tcp port, and
a random UUIDesque id.

The negotiation process is also changed a bit now; we check for an OK_NEGOTIATE
from all peers before sending QUERY_CAPS.