[MAJOR CHANGE] OvLocalPeer is now a GObject derived type
This is essential for us since we now need signals for various events happening
to the OvLocalPeer in order to drive the GUI.

Currently there's no signals anywhere and this change just brings the new code
to feature-parity with the old code.

I refrained from porting OvRemotePeer and OvDiscoveredPeer to GObject as well
since that would take up a lot more time for no current benefit. If we need
signals for those as well, I will port those over to GObject too. They will
simply be derived types of OvPeer.