Add signals for negotiation and calling for use with the GUI
This allows us to handle incoming calls from within a GUI, handle aborted
negotiations, ended calls, etc. The earlier method of checking the state was
obviously broken and was a placeholder for this method.

I intended to have this done by a series of smaller commits, but there were
enough problems and yak-shavings involved that it became rather difficult and
a waste of time to try and split the working code into a series of commits.