ov: Implement timing out of remotes during a call
If a remote peer crashes or loses network connectivity, we signal the
application about this after the timeout is reached with the "call-remote-gone"
signal. If all remotes go away, the "call-all-remotes-gone" signal is emitted;
at which point the application should hang-up the call.

TODO: The mechanism surrounding the "call-remote-gone" signal has not been fully
tested yet.