ov: Rewrite code responsible for caps negotiation and RTP/RTCP port exchange
The old code was one massive impossible-to-understand nested loop. That has now
been split out into several loops and intermediatory datastructures that make
the whole process easier to understand (I hope). This was done because there was
a bug in the caps negotiation where the wrong caps were getting assigned to
peers and it was easier to rewrite the loops than to fix the earlier code.

Also use unsigned 16-bit integers for ports everywhere which makes the API
everywhere more consistent (no up/down casting required).